Refactor: ES5
This PR refactors the code by putting it into a single file and using
plain ES5 rather than ES2015 and Babel.
