--- .babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": [
-    "es2015"
-  ],
-  "plugins": [
-    "add-module-exports"
-  ]
-}
