+++ .gitignore
@@ -2,4 +2,3 @@
 /node_modules/
 /typings/
 /npm-debug.log
-/chai-iterator.js
