+++ package.json
@@ -5,21 +5,14 @@
   "main": "chai-iterator.js",
   "typings": "chai-iterator.d.ts",
   "scripts": {
-    "build": "browserify -e src/index.js -o chai-iterator.js -s chai-iterator --bare",
-    "unbuild": "rimraf chai-iterator.js",
     "test:tsc": "tsc",
-    "test:eslint": "eslint src",
+    "test:eslint": "eslint chai-iterator.js",
     "test:mocha": "mocha",
     "test:cover": "istanbul cover _mocha",
     "test:report": "npm run test:cover && open coverage/lcov-report/index.html",
     "test": "npm run test:tsc && npm run test:eslint && npm run test:cover",
     "typings": "typings install",
     "coveralls": "cat coverage/lcov.info | coveralls"
-  },
-  "browserify": {
-    "transform": [
-      "babelify"
-    ]
   },
   "repository": {
     "type": "git",
@@ -52,11 +45,6 @@
     "chai": "3.x"
   },
   "devDependencies": {
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-register": "^6.9.0",
-    "babelify": "^7.3.0",
-    "browserify": "^13.0.1",
     "chai": "^3.5.0",
     "coffee-script": "^1.10.0",
     "core-js": "^2.4.0",
