--- src/index.js
@@ -1,6 +1,0 @@
-import chaiIterator from './chai-iterator';
-import setup from './setup';
-
-setup(chaiIterator);
-
-export default chaiIterator;
