--- src/util/is-iterable.js
@@ -1,3 +1,0 @@
-export default function(value) {
-  return value != null && typeof value[Symbol.iterator] === 'function';
-}
