--- src/util/unquote.js
@@ -1,17 +1,0 @@
-export default function(string) {
-  return new Unquote(string);
-}
-
-class Unquote {
-  constructor(str) {
-    this._str = str;
-  }
-
-  inspect() {
-    return this.toString();
-  }
-
-  toString() {
-    return this._str;
-  }
-}
