+++ test/mocha.opts
@@ -1,3 +1,3 @@
 --recursive
 --reporter dot
---compilers js:babel-register,coffee:coffee-script/register
+--compilers coffee:coffee-script/register
